refactor(chatroom): ChatRoom.jsx 분리 리팩토링
- src/utils/chatFormat.js: 시간/날짜 포맷 유틸 3개 분리
- ChatMessageItem.jsx: 메시지 단일 렌더링 컴포넌트 생성
- ChatInputBar.jsx: 입력창 + 전송 로직 자체 관리 컴포넌트 생성
- ChatContextMenu.jsx: 메시지 컨텍스트 메뉴 컴포넌트 생성
- ChatRoom.jsx: 909줄 → 377줄 (59% 감소)
