+++ src/pages/Chat/ChatRoom.jsx
@@ -5,10 +5,6 @@ import styled from 'styled-components';
 import AlertModal from '../../components/common/AlertModal';
 import BottomModal from '../../components/common/BottomModal';
 import Header from '../../components/common/Header';
-import ImageIcon from '../../assets/icons/icon-image.svg?react';
-import HeartIcon from '../../assets/icons/icon-heart.svg?react';
-import EmojiIcon from '../../assets/icons/icon-emoji.svg?react';
-import EmojiPicker, { STICKER_MAP } from '../../components/chat/EmojiPicker';
 import heartFillSrc from '../../assets/emoji/icon_heart_fill.svg';
 import thumbsUpSrc from '../../assets/emoji/icon_thumbs_up_fill.png';
 import starSrc from '../../assets/emoji/icon_star.png';
@@ -16,9 +12,6 @@ import { useAuth } from '../../context/AuthContext';
 import { db } from '../../firebase/config';
 import {
   subscribeToMessages,
-  sendTextMessage,
-  sendImageMessage,
-  sendStickerMessage,
   toggleReaction,
   markAsRead,
   editMessage,
@@ -28,15 +21,17 @@ import {
 } from '../../firebase/chat';
 import { uploadImage } from '../../api/auth';
 import { getImageUrl } from '../../utils/format';
-import Avatar from '../../components/common/Avatar';
 import ChatThemePanel, { BG_COLORS, BUBBLE_COLORS } from './ChatThemePanel';
 import InviteUserModal from '../../components/chat/InviteUserModal';
+import ChatMessageItem from '../../components/chat/ChatMessageItem';
+import ChatInputBar from '../../components/chat/ChatInputBar';
+import ChatContextMenu from '../../components/chat/ChatContextMenu';
 
-const REACTION_TYPES = [
-  { key: 'heart', src: heartFillSrc },
-  { key: 'thumbs_up', src: thumbsUpSrc },
-  { key: 'star', src: starSrc },
-];
+const REACTION_SRC_MAP = {
+  heart: heartFillSrc,
+  thumbs_up: thumbsUpSrc,
+  star: starSrc,
+};
 
 const Wrapper = styled.div`
   min-height: 100vh;
@@ -58,328 +53,17 @@ const MessageList = styled.div`
   gap: 12px;
 `;
 
-const DateDivider = styled.div`
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  margin: 8px 0 4px;
-
-  &::before,
-  &::after {
-    content: '';
-    flex: 1;
-    height: 1px;
-    background-color: ${({ theme }) => theme.colors.border};
-  }
-`;
-
-const DateDividerText = styled.span`
-  font-size: ${({ theme }) => theme.fonts.size.xs};
-  color: ${({ theme }) => theme.colors.gray400};
-  background-color: ${({ theme }) => theme.colors.white};
-  border: 1px solid ${({ theme }) => theme.colors.border};
-  border-radius: ${({ theme }) => theme.borderRadius.round};
-  padding: 4px 10px;
-`;
-
-const MessageWrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: ${({ $isMine }) => ($isMine ? 'flex-end' : 'flex-start')};
-`;
-
-const MessageRow = styled.div`
-  display: flex;
-  align-items: flex-end;
-  gap: 8px;
-  width: 100%;
-  flex-direction: ${({ $isMine }) => ($isMine ? 'row-reverse' : 'row')};
-`;
-
-const Bubble = styled.div`
-  max-width: 60%;
-  padding: 10px 14px;
-  border-radius: ${({ $isMine }) => ($isMine ? '16px 0 16px 16px' : '0 16px 16px 16px')};
-  background-color: ${({ $isMine, $bubbleColor, $otherBubbleColor, theme }) =>
-    $isMine ? $bubbleColor || theme.colors.primary : $otherBubbleColor || theme.colors.white};
-  color: ${({ $isMine, $otherBubbleColor, theme }) =>
-    $isMine || $otherBubbleColor ? theme.colors.white : theme.colors.black};
-  font-size: ${({ theme }) => theme.fonts.size.base};
-  line-height: 1.5;
-  word-break: break-word;
-`;
-
-const ChatTime = styled.span`
-  font-size: ${({ theme }) => theme.fonts.size.xs};
-  color: ${({ theme }) => theme.colors.gray400};
-`;
-
-const ChatImage = styled.img`
-  max-width: 200px;
-  border-radius: ${({ theme }) => theme.borderRadius.base};
-  object-fit: cover;
-`;
-
-const StickerImg = styled.img`
-  width: 120px;
-  height: 120px;
-  object-fit: contain;
-`;
-
-const InputArea = styled.div`
-  position: fixed;
-  bottom: 0;
-  left: 50%;
-  transform: translateX(-50%);
-  width: 100%;
-  max-width: 390px;
-  background-color: ${({ theme }) => theme.colors.white};
-  border-top: 1px solid ${({ theme }) => theme.colors.border};
-  display: flex;
-  align-items: center;
-  padding: 10px 16px;
-  gap: 8px;
-`;
-
-const ImageInputBtn = styled.button`
-  width: 36px;
-  height: 36px;
-  background-color: ${({ theme }) => theme.colors.gray200};
-  border: none;
-  outline: none;
-  border-radius: ${({ theme }) => theme.borderRadius.round};
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  flex-shrink: 0;
-  opacity: ${({ disabled }) => (disabled ? 0.4 : 1)};
-`;
-
-const EmojiBtn = styled.button`
-  width: 36px;
-  height: 36px;
-  background-color: ${({ $active, theme }) => ($active ? theme.colors.gray300 : theme.colors.gray200)};
-  border: none;
-  outline: none;
-  border-radius: ${({ theme }) => theme.borderRadius.round};
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  flex-shrink: 0;
-  transition: background 0.1s;
-`;
-
-const TextInput = styled.input`
-  flex: 1;
-  background-color: ${({ theme }) => theme.colors.gray100};
-  border: none;
-  outline: none;
-  border-radius: ${({ theme }) => theme.borderRadius.round};
-  padding: 8px 16px;
-  font-size: ${({ theme }) => theme.fonts.size.base};
-  color: ${({ theme }) => theme.colors.black};
-
-  &::placeholder {
-    color: ${({ theme }) => theme.colors.gray300};
-  }
-`;
-
-const SendButton = styled.button`
-  border: none;
-  outline: none;
-  background: transparent;
-  font-size: ${({ theme }) => theme.fonts.size.sm};
-  font-weight: ${({ theme }) => theme.fonts.weight.medium};
-  color: ${({ disabled, theme }) => (disabled ? theme.colors.gray300 : theme.colors.primary)};
-`;
-
-const ContextMenu = styled.ul`
-  position: fixed;
-  background: ${({ theme }) => theme.colors.white};
-  border: 1px solid ${({ theme }) => theme.colors.border};
-  border-radius: ${({ theme }) => theme.borderRadius.base};
-  box-shadow: ${({ theme }) => theme.shadows.base};
-  z-index: 250;
-  list-style: none;
-  padding: 4px 0;
-  min-width: 120px;
-`;
-
-const ContextMenuItem = styled.li`
-  padding: 10px 16px;
-  font-size: ${({ theme }) => theme.fonts.size.sm};
-  cursor: pointer;
-  text-align: center;
-  color: ${({ $danger, theme }) => ($danger ? theme.colors.error : theme.colors.text)};
-  &:hover {
-    background: ${({ theme }) => theme.colors.gray100};
-  }
-`;
-
-const ReactionRow = styled.li`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 8px;
-  padding: 8px 12px;
-  border-bottom: 1px solid ${({ theme }) => theme.colors.border};
-`;
-
-const ReactionBtn = styled.button`
-  width: 36px;
-  height: 36px;
-  border: none;
-  border-radius: 50%;
-  background: ${({ $active, theme }) => ($active ? theme.colors.gray200 : 'transparent')};
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  cursor: pointer;
-  transition: background 0.1s;
-
-  &:hover {
-    background: ${({ theme }) => theme.colors.gray100};
-  }
-
-  img {
-    width: 22px;
-    height: 22px;
-    object-fit: contain;
-  }
-`;
-
-const EditWrapper = styled.div`
-  position: relative;
-  max-width: 60%;
-`;
-
-const EditInput = styled.input`
-  width: 100%;
-  padding: 10px 52px 10px 14px;
-  border-radius: ${({ theme }) => theme.borderRadius.base};
-  border: 1.5px solid ${({ theme }) => theme.colors.primary};
-  font-size: ${({ theme }) => theme.fonts.size.base};
-  outline: none;
-  box-sizing: border-box;
-`;
-
-const HeartReaction = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: ${({ $isMine }) => ($isMine ? 'flex-end' : 'flex-start')};
-  padding: ${({ $isMine }) => ($isMine ? '4px 8px 0 0' : '4px 0 0 40px')};
-`;
-
-const HeartBubble = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 20px;
-  height: 20px;
-  background-color: #e53935;
-  border: none;
-  border-radius: 50%;
-  box-shadow: ${({ theme }) => theme.shadows.base};
-
-  svg {
-    width: 10px;
-    height: 10px;
-    path {
-      fill: #ffffff;
-      stroke: #ffffff;
-    }
-  }
-`;
-
-const ReactionBar = styled.div`
-  display: flex;
-  align-items: center;
-  flex-wrap: wrap;
-  gap: 4px;
-  max-width: 60%;
-  padding: ${({ $isMine }) => ($isMine ? '4px 8px 0 0' : '4px 0 0 40px')};
-`;
-
-const ReactionPill = styled.div`
-  display: flex;
-  align-items: center;
-  gap: 3px;
-  background: ${({ theme }) => theme.colors.white};
-  border: 1px solid ${({ theme }) => theme.colors.border};
-  border-radius: ${({ theme }) => theme.borderRadius.round};
-  padding: 2px 6px;
-  box-shadow: 0px 1px 3px rgba(0, 0, 0, 0.08);
-  cursor: pointer;
-
-  img {
-    width: 14px;
-    height: 14px;
-    object-fit: contain;
-  }
-
-  span {
-    font-size: 11px;
-    color: ${({ theme }) => theme.colors.gray500};
-  }
-`;
-
-const EditConfirmBtn = styled.button`
-  position: absolute;
-  right: 8px;
-  top: 50%;
-  transform: translateY(-50%);
-  padding: 4px 8px;
-  border: none;
-  border-radius: ${({ theme }) => theme.borderRadius.sm};
-  background-color: ${({ theme }) => theme.colors.primary};
-  color: ${({ theme }) => theme.colors.white};
-  font-size: ${({ theme }) => theme.fonts.size.xs};
-  font-weight: ${({ theme }) => theme.fonts.weight.medium};
-  cursor: pointer;
-`;
-
-const formatMsgTime = (timestamp) => {
-  if (!timestamp) return '';
-  const date = timestamp.toDate();
-  return date.toLocaleTimeString('ko-KR', { hour: '2-digit', minute: '2-digit', hour12: false });
-};
-
-const getMsgDateKey = (timestamp) => {
-  if (!timestamp) return '';
-  const date = timestamp.toDate();
-  const y = date.getFullYear();
-  const m = String(date.getMonth() + 1).padStart(2, '0');
-  const d = String(date.getDate()).padStart(2, '0');
-  return `${y}-${m}-${d}`;
-};
-
-const formatMsgDate = (timestamp) => {
-  if (!timestamp) return '';
-  const date = timestamp.toDate();
-  return date.toLocaleDateString('ko-KR', {
-    year: 'numeric',
-    month: 'long',
-    day: 'numeric',
-    weekday: 'short',
-  });
-};
-
 const ChatRoom = () => {
   const { chatId } = useParams();
   const navigate = useNavigate();
   const { user } = useAuth();
-  const fileRef = useRef(null);
   const bottomRef = useRef(null);
   const contextMenuRef = useRef(null);
-  const inputContextRef = useRef({ selectionStart: 0, selectionEnd: 0 });
 
   const [chatInfo, setChatInfo] = useState(null);
   const [messages, setMessages] = useState([]);
-  const [inputText, setInputText] = useState('');
-  const [isSending, setIsSending] = useState(false);
   const [showModal, setShowModal] = useState(false);
   const [showInviteModal, setShowInviteModal] = useState(false);
-  const [showEmojiPicker, setShowEmojiPicker] = useState(false);
   const [contextMenu, setContextMenu] = useState({
     show: false,
     anchorRect: null,
@@ -450,13 +134,6 @@ const ChatRoom = () => {
     return () => window.removeEventListener('click', handleClick);
   }, [contextMenu.show]);
 
-  useEffect(() => {
-    if (!showEmojiPicker) return;
-    const handleClick = () => setShowEmojiPicker(false);
-    window.addEventListener('click', handleClick);
-    return () => window.removeEventListener('click', handleClick);
-  }, [showEmojiPicker]);
-
   useLayoutEffect(() => {
     if (!contextMenu.show || !contextMenuRef.current || !contextMenu.anchorRect) return;
     const menu = contextMenuRef.current;
@@ -465,18 +142,8 @@ const ChatRoom = () => {
     const vw = window.innerWidth;
     const vh = window.innerHeight;
 
-    let x, y;
-
-    if (contextMenu.type === 'input') {
-      x = anchorRect.left;
-      y = anchorRect.top - height;
-    } else if (contextMenu.isMine) {
-      x = anchorRect.left - width;
-      y = anchorRect.top;
-    } else {
-      x = anchorRect.right;
-      y = anchorRect.top;
-    }
+    let x = contextMenu.isMine ? anchorRect.left - width : anchorRect.right;
+    let y = anchorRect.top;
 
     if (x + width > vw) x = vw - width - 8;
     if (y + height > vh) y = vh - height - 8;
@@ -485,7 +152,7 @@ const ChatRoom = () => {
 
     menu.style.left = `${x}px`;
     menu.style.top = `${y}px`;
-  }, [contextMenu.show, contextMenu.anchorRect, contextMenu.type, contextMenu.isMine]);
+  }, [contextMenu.show, contextMenu.anchorRect, contextMenu.isMine]);
 
   const otherParticipant = (() => {
     if (!chatInfo || !user?.accountname) return null;
@@ -493,48 +160,6 @@ const ChatRoom = () => {
     const info = chatInfo.participantInfo?.[otherAccountname] || { username: '', image: '' };
     return { ...info, accountname: otherAccountname };
   })();
-
-  const handleSend = async () => {
-    const text = inputText.trim();
-    if (!text || isSending) return;
-    setInputText('');
-    setIsSending(true);
-    try {
-      await sendTextMessage(chatId, user.accountname, text);
-    } finally {
-      setIsSending(false);
-    }
-  };
-
-  const handleKeyDown = (e) => {
-    if (e.key === 'Enter' && !e.shiftKey) {
-      e.preventDefault();
-      handleSend();
-    }
-  };
-
-  const handleImageUpload = async (e) => {
-    const file = e.target.files?.[0];
-    if (!file) return;
-    setIsSending(true);
-    try {
-      await sendImageMessage(chatId, user.accountname, file);
-    } finally {
-      setIsSending(false);
-      e.target.value = '';
-    }
-  };
-
-  const handleStickerSend = async (stickerKey) => {
-    setShowEmojiPicker(false);
-    if (isSending) return;
-    setIsSending(true);
-    try {
-      await sendStickerMessage(chatId, user.accountname, stickerKey);
-    } finally {
-      setIsSending(false);
-    }
-  };
 
   const handleContextMenu = (e, msg, isMine) => {
     e.preventDefault();
@@ -563,12 +188,6 @@ const ChatRoom = () => {
     setEditingId(null);
   };
 
-  const handleEditSubmit = async (e, msg) => {
-    if (e.key !== 'Enter' || e.shiftKey) return;
-    e.preventDefault();
-    await confirmEdit(msg);
-  };
-
   const handleDelete = () => {
     setContextMenu((prev) => ({ ...prev, show: false }));
     setShowDeleteMsgAlert(true);
@@ -589,35 +208,6 @@ const ChatRoom = () => {
   const handleReport = () => {
     setContextMenu((prev) => ({ ...prev, show: false }));
     setShowReportAlert(true);
-  };
-
-  const handleInputContextMenu = (e) => {
-    e.preventDefault();
-    inputContextRef.current = {
-      selectionStart: e.target.selectionStart,
-      selectionEnd: e.target.selectionEnd,
-    };
-    const rect = e.currentTarget.getBoundingClientRect();
-    setContextMenu({
-      show: true,
-      anchorRect: { left: rect.left, right: rect.right, top: rect.top, bottom: rect.bottom },
-      messageId: null,
-      isMine: false,
-      text: '',
-      type: 'input',
-      reactions: {},
-    });
-  };
-
-  const handlePaste = async () => {
-    try {
-      const clipText = await navigator.clipboard.readText();
-      const { selectionStart, selectionEnd } = inputContextRef.current;
-      setInputText((prev) => prev.slice(0, selectionStart) + clipText + prev.slice(selectionEnd));
-    } catch {
-      // 클립보드 권한 없을 경우 무시
-    }
-    setContextMenu((prev) => ({ ...prev, show: false }));
   };
 
   const handleBgImageChange = async (file) => {
@@ -675,129 +265,31 @@ const ChatRoom = () => {
         />
 
         <MessageList style={{ visibility: themeReady ? 'visible' : 'hidden' }}>
-          {messages.map((msg, index) => {
-            const isMine = msg.senderId === user?.accountname;
-            const currentDateKey = getMsgDateKey(msg.createdAt);
-            const prevDateKey = index > 0 ? getMsgDateKey(messages[index - 1].createdAt) : '';
-            const showDateDivider = index === 0 || currentDateKey !== prevDateKey;
-            const currentTime = formatMsgTime(msg.createdAt);
-            const nextMsg = messages[index + 1];
-            const nextDateKey = nextMsg ? getMsgDateKey(nextMsg.createdAt) : '';
-            const nextTime = nextMsg ? formatMsgTime(nextMsg.createdAt) : '';
-            const showTime =
-              !nextMsg ||
-              nextDateKey !== currentDateKey ||
-              nextMsg.senderId !== msg.senderId ||
-              nextTime !== currentTime;
-
-            const hasAnyReaction = REACTION_TYPES.some(({ key }) => (msg.reactions?.[key]?.length || 0) > 0);
-
-            return (
-              <div key={msg.id}>
-                {showDateDivider && (
-                  <DateDivider>
-                    <DateDividerText>{formatMsgDate(msg.createdAt)}</DateDividerText>
-                  </DateDivider>
-                )}
-
-                <MessageWrapper $isMine={isMine}>
-                  <MessageRow $isMine={isMine}>
-                    {!isMine && (
-                      <Avatar
-                        src={chatInfo?.participantInfo?.[msg.senderId]?.image || otherParticipant?.image}
-                        alt="상대방"
-                        size="32px"
-                        onClick={() => navigate(`/profile/${msg.senderId}`)}
-                      />
-                    )}
-                    {msg.stickerKey ? (
-                      <StickerImg
-                        src={STICKER_MAP[msg.stickerKey]}
-                        alt="스티커"
-                        onContextMenu={(e) => handleContextMenu(e, msg, isMine)}
-                      />
-                    ) : msg.imageUrl ? (
-                      <ChatImage src={msg.imageUrl} alt="채팅 이미지" />
-                    ) : editingId === msg.id ? (
-                      <EditWrapper>
-                        <EditInput
-                          autoFocus
-                          value={editText}
-                          onChange={(e) => setEditText(e.target.value)}
-                          onKeyDown={(e) => handleEditSubmit(e, msg)}
-                        />
-                        <EditConfirmBtn onMouseDown={(e) => e.preventDefault()} onClick={() => confirmEdit(msg)}>
-                          완료
-                        </EditConfirmBtn>
-                      </EditWrapper>
-                    ) : (
-                      <Bubble
-                        $isMine={isMine}
-                        $bubbleColor={bubbleColor}
-                        $otherBubbleColor={otherBubbleColor}
-                        onContextMenu={(e) => handleContextMenu(e, msg, isMine)}
-                      >
-                        {msg.text}
-                      </Bubble>
-                    )}
-                    {showTime && <ChatTime>{currentTime}</ChatTime>}
-                  </MessageRow>
-
-                  {hasAnyReaction && (
-                    <ReactionBar $isMine={isMine}>
-                      {REACTION_TYPES.filter(({ key }) => (msg.reactions?.[key]?.length || 0) > 0).map(
-                        ({ key, src }) => {
-                          const hasReacted = msg.reactions[key]?.includes(user.accountname) || false;
-                          return (
-                            <ReactionPill
-                              key={key}
-                              onClick={() => toggleReaction(chatId, msg.id, user.accountname, key, hasReacted)}
-                            >
-                              <img src={src} alt={key} />
-                              <span>{msg.reactions[key].length}</span>
-                            </ReactionPill>
-                          );
-                        },
-                      )}
-                    </ReactionBar>
-                  )}
-                </MessageWrapper>
-              </div>
-            );
-          })}
+          {messages.map((msg, index) => (
+            <ChatMessageItem
+              key={msg.id}
+              msg={msg}
+              prevMsg={index > 0 ? messages[index - 1] : null}
+              nextMsg={messages[index + 1] || null}
+              isMine={msg.senderId === user?.accountname}
+              bubbleColor={bubbleColor}
+              otherBubbleColor={otherBubbleColor}
+              editingId={editingId}
+              editText={editText}
+              setEditText={setEditText}
+              onConfirmEdit={confirmEdit}
+              onContextMenu={handleContextMenu}
+              chatInfo={chatInfo}
+              user={user}
+              reactionSrcMap={REACTION_SRC_MAP}
+              chatId={chatId}
+            />
+          ))}
           <div ref={bottomRef} />
         </MessageList>
       </Wrapper>
 
-      <EmojiPicker isOpen={showEmojiPicker} onSelect={handleStickerSend} />
-
-      <InputArea>
-        <ImageInputBtn onClick={() => fileRef.current?.click()} disabled={isSending}>
-          <ImageIcon width="22" height="22" />
-        </ImageInputBtn>
-        <input ref={fileRef} type="file" accept="image/*" style={{ display: 'none' }} onChange={handleImageUpload} />
-        <EmojiBtn
-          $active={showEmojiPicker}
-          onClick={(e) => {
-            e.stopPropagation();
-            setShowEmojiPicker((prev) => !prev);
-          }}
-          disabled={isSending}
-        >
-          <EmojiIcon width="22" height="22" color="white" />
-        </EmojiBtn>
-        <TextInput
-          value={inputText}
-          onChange={(e) => setInputText(e.target.value)}
-          onKeyDown={handleKeyDown}
-          onContextMenu={handleInputContextMenu}
-          onFocus={() => setShowEmojiPicker(false)}
-          placeholder="메시지 입력하기..."
-        />
-        <SendButton onClick={handleSend} disabled={!inputText.trim() || isSending}>
-          전송
-        </SendButton>
-      </InputArea>
+      <ChatInputBar chatId={chatId} senderAccountname={user?.accountname} />
 
       <BottomModal isOpen={showModal} onClose={() => setShowModal(false)} items={modalItems} />
 
@@ -867,41 +359,17 @@ const ChatRoom = () => {
         existingParticipants={chatInfo?.participants || []}
       />
 
-      {contextMenu.show && (
-        <ContextMenu ref={contextMenuRef} style={{ top: 0, left: 0 }} onClick={(e) => e.stopPropagation()}>
-          {contextMenu.type === 'input' ? (
-            <ContextMenuItem onClick={handlePaste}>붙여넣기</ContextMenuItem>
-          ) : (
-            <>
-              <ReactionRow>
-                {REACTION_TYPES.map(({ key, src }) => {
-                  const hasReacted = contextMenu.reactions?.[key]?.includes(user.accountname) || false;
-                  return (
-                    <ReactionBtn key={key} $active={hasReacted} onClick={() => handleReaction(key)}>
-                      <img src={src} alt={key} />
-                    </ReactionBtn>
-                  );
-                })}
-              </ReactionRow>
-              {contextMenu.isMine ? (
-                <>
-                  {contextMenu.text && <ContextMenuItem onClick={handleEditStart}>수정</ContextMenuItem>}
-                  <ContextMenuItem $danger onClick={handleDelete}>
-                    삭제
-                  </ContextMenuItem>
-                </>
-              ) : (
-                <>
-                  {contextMenu.text && <ContextMenuItem onClick={handleCopy}>복사</ContextMenuItem>}
-                  <ContextMenuItem $danger onClick={handleReport}>
-                    신고
-                  </ContextMenuItem>
-                </>
-              )}
-            </>
-          )}
-        </ContextMenu>
-      )}
+      <ChatContextMenu
+        contextMenu={contextMenu}
+        contextMenuRef={contextMenuRef}
+        user={user}
+        reactionSrcMap={REACTION_SRC_MAP}
+        onReaction={handleReaction}
+        onEditStart={handleEditStart}
+        onDelete={handleDelete}
+        onCopy={handleCopy}
+        onReport={handleReport}
+      />
     </>
   );
 };
